Require Tokio for stream consumer
Requiring Tokio admits a much simpler implementation that does not
require a separate thread or executor. The dependency on Tokio is
optional (on by default) in order to allow folks who are using only the
BaseConsumer to avoid the Tokio dependency.

Fix #192.
