+++ src/lib.rs
@@ -1,8 +1,9 @@
 //! # rust-rdkafka
+//!
 //! A fully asynchronous, [futures]-based Kafka client library for Rust based on [librdkafka].
 //!
 //! ## The library
-//! `rust-rdkafka` provides a safe Rust interface to librdkafka. The master branch is currently based on librdkafka 0.11.6.
+//! `rust-rdkafka` provides a safe Rust interface to librdkafka. The master branch is currently based on librdkafka 1.3.0.
 //!
 //! ### Documentation
 //!
@@ -55,24 +56,14 @@
 //!
 //! For more information about consumers and producers, refer to their module-level documentation.
 //!
-//! [`BaseConsumer`]: consumer/base_consumer/struct.BaseConsumer.html
-//! [`BaseProducer`]: producer/base_producer/struct.BaseProducer.html
-//! [`ThreadedProducer`]: producer/base_producer/struct.ThreadedProducer.html
-//! [`StreamConsumer`]: consumer/stream_consumer/struct.StreamConsumer.html
-//! [`FutureProducer`]: producer/future_producer/struct.FutureProducer.html
-//! [librdkafka]: https://github.com/edenhill/librdkafka
-//! [futures]: https://github.com/alexcrichton/futures-rs
-//! [`future`]: https://docs.rs/futures/0.1.3/futures/trait.Future.html
-//! [`stream`]: https://docs.rs/futures/0.1.3/futures/stream/trait.Stream.html
-//!
 //! *Warning*: the library is under active development and the APIs are likely to change.
 //!
-//! ### Asynchronous data processing with tokio-rs
-//! [tokio-rs] is a platform for fast processing of asynchronous events in Rust. The interfaces exposed by the [`StreamConsumer`] and the [`FutureProducer`] allow rust-rdkafka users to easily integrate Kafka consumers and producers within the tokio-rs platform, and write asynchronous message processing code. Note that rust-rdkafka can be used without tokio-rs.
+//! ### Asynchronous data processing with Tokio
+//! [Tokio] is a platform for fast processing of asynchronous events in Rust. The interfaces exposed by the [`StreamConsumer`] and the [`FutureProducer`] allow rust-rdkafka users to easily integrate Kafka consumers and producers within the Tokio platform, and write asynchronous message processing code. Note that rust-rdkafka can be used without Tokio.
 //!
-//! To see rust-rdkafka in action with tokio-rs, check out the [asynchronous processing example] in the examples folder.
+//! To see rust-rdkafka in action with Tokio, check out the [asynchronous processing example] in the examples folder.
 //!
-//! [tokio-rs]: https://tokio.rs/
+//! [Tokio]: https://tokio.rs/
 //! [asynchronous processing example]: https://github.com/fede1024/rust-rdkafka/blob/master/examples/asynchronous_processing.rs
 //!
 //! ### At-least-once delivery
@@ -107,33 +98,52 @@
 //!
 //! ```toml
 //! [dependencies]
-//! rdkafka = "~0.21"
+//! rdkafka = { version = "0.23", features = ["cmake-build"] }
 //! ```
 //!
-//! This crate will compile librdkafka from sources and link it statically to your executable. To compile librdkafka you'll need:
+//! This crate will compile librdkafka from sources and link it statically to your
+//! executable. To compile librdkafka you'll need:
 //!
 //! * the GNU toolchain
 //! * GNU `make`
 //! * `pthreads`
-//! * `zlib`
-//! * `libssl-dev`: optional, *not* included by default (feature: `ssl`).
-//! * `libsasl2-dev`: optional, *not* included by default (feature: `sasl`).
+//! * `zlib`: optional, but included by default (feature: `libz`)
+//! * `cmake`: optional, *not* included by default (feature: `cmake-build`)
+//! * `libssl-dev`: optional, *not* included by default (feature: `ssl`)
+//! * `libsasl2-dev`: optional, *not* included by default (feature: `gssapi`)
+//! * `libzstd-dev`: optional, *not* included by default (feature: `zstd-pkg-config`)
 //!
-//! To enable ssl and sasl, use the `features` field in `Cargo.toml`. Example:
+//! Note that using the CMake build system, via the `cmake-build` feature, is
+//! **strongly** encouraged. The default build system has a [known
+//! issue](rdkafka-sys/README.md#known-issues) that can cause corrupted builds.
+//!
+//! By default a submodule with the librdkafka sources pinned to a specific commit
+//! will be used to compile and statically link the library. The `dynamic-linking`
+//! feature can be used to instead dynamically link rdkafka to the system's version
+//! of librdkafka. Example:
 //!
 //! ```toml
-//! [dependencies.rdkafka]
-//! version = "~0.21"
-//! features = ["ssl", "sasl"]
+//! [dependencies]
+//! rdkafka = { version = "0.23", features = ["dynamic-linking"] }
 //! ```
 //!
-//! By default a submodule with the librdkafka sources pinned to a specific commit will
-//! be used to compile and statically link the library.
+//! For a full listing of features, consult the [rdkafka-sys crate's
+//! documentation](rdkafka-sys/README.md#features). All of rdkafka-sys features are
+//! re-exported as rdkafka features.
 //!
-//! The `dynamic_linking` feature can be used to link rdkafka to a locally installed
-//! version of librdkafka: if the feature is enabled, the build script will use `pkg-config`
-//! to check the version of the library installed in the system, and it will configure the
-//! compiler to use dynamic linking.
+//! ### Tokio dependency
+//!
+//! The [`StreamConsumer`] and [`FutureProducer`] depend on Tokio, which can be a
+//! heavyweight dependency for users who only intend to use the non-async/await
+//! consumers or producers.
+//!
+//! The `tokio` feature is enabled by default. To disable it, turn off default
+//! features like so:
+//!
+//! ```toml
+//! [dependencies]
+//! rdkafka = { version = "0.23", default-features = false }
+//! ```
 //!
 //! ## Compiling from sources
 //!
@@ -146,7 +156,7 @@
 //! and then compile using `cargo`, selecting the features that you want. Example:
 //!
 //! ```bash
-//! cargo build --features "ssl sasl"
+//! cargo build --features "ssl gssapi"
 //! ```
 //!
 //! ## Examples
@@ -206,8 +216,18 @@
 //! To enable debugging in your project, make sure you initialize the logger with
 //! `env_logger::init()` or equivalent.
 //!
+//! [`BaseConsumer`]: https://docs.rs/rdkafka/*/rdkafka/consumer/base_consumer/struct.BaseConsumer.html
+//! [`BaseProducer`]: https://docs.rs/rdkafka/*/rdkafka/producer/base_producer/struct.BaseProducer.html
+//! [`ThreadedProducer`]: https://docs.rs/rdkafka/*/rdkafka/producer/base_producer/struct.ThreadedProducer.html
+//! [`StreamConsumer`]: https://docs.rs/rdkafka/*/rdkafka/consumer/stream_consumer/struct.StreamConsumer.html
+//! [`FutureProducer`]: https://docs.rs/rdkafka/*/rdkafka/producer/future_producer/struct.FutureProducer.html
+//! [librdkafka]: https://github.com/edenhill/librdkafka
+//! [futures]: https://github.com/alexcrichton/futures-rs
+//! [`future`]: https://docs.rs/futures/0.1.3/futures/trait.Future.html
+//! [`stream`]: https://docs.rs/futures/0.1.3/futures/stream/trait.Stream.html
 
 #![warn(missing_docs)]
+#![cfg_attr(docsrs, feature(doc_cfg))]
 
 pub use rdkafka_sys::types;
